+++ ximpol/irf/mrf.py
@@ -20,7 +20,7 @@
 import numpy
 from astropy.io import fits
 
-from ximpol.utils.logging_ import logger
+from ximpol.utils.logging_ import logger, abort
 from ximpol.irf.base import OGIP_HEADER_SPECS
 from ximpol.core.fitsio import xBinTableHDUBase
 from ximpol.core.spline import xInterpolatedUnivariateSplineLinear
@@ -489,6 +489,17 @@ class xModulationFactor(xInterpolatedUnivariateSplineLinear):
             The polarization angle, in radians. (This can either be a vector or
             an array of the same length as `energy`.)
         """
+        try:
+            min_degree = polarization_degree.min()
+            max_degree = polarization_degree.max()
+        except AttributeError:
+            # This is catching the case where the polarization degree is
+            # constant and is passed through as a float.
+            min_degree = max_degree = polarization_degree
+        if max_degree > 1:
+            abort('The polarization degree must be <= 1')
+        if min_degree < 0:
+            abort('The polarization degree must be >= 0')
         visibility = self(energy)*polarization_degree
         return self.generator.rvs_phi(visibility, polarization_angle)